Weather Fixed

+++ src/main/java/org/example/Models/Player.java
@@ -3,7 +3,7 @@ package org.example.Models;
 public class Player {
     public int PositionX;
     public int PositionY;
-    public int Energy=200;
+    public double Energy=200;
     public boolean Fainted=false;
     public int FarmingSkill=0;
     public int MiningSkill=0;
